+++ src/package/themes/default/components/splitter.ts
@@ -1,0 +1,11 @@
+export default {
+    borderColor: 'var(--vf-border-color)',
+    borderRadius: 'var(--vf-radii-md)',
+    panelBackgroundColor: 'var(--vf-controls-background-color)',
+    handleWidth: '3px',
+    handleHeight: '1.5rem',
+    handleRadius: '999px',
+    handleColor: 'var(--vf-border-color)',
+    gutterActiveBackgroundColor: 'rgba(var(--vf-blue-600-rgb), 0.12)',
+    disabledOpacity: 'var(--vf-states-disabled-opacity)',
+};